service/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
click==6.7
Flask==1.0.2
Flask-PyMongo==2.1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pymongo==3.7.1
Werkzeug==0.14.1
click==6.7
Flask==1.0.2
Flask-PyMongo==2.1.0
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pymongo==3.7.1
Werkzeug==0.14.1
click==6.7
Flask==1.0.2
Flask-PyMongo==2.1.0
functools32==3.2.3.post2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pymongo==3.7.1
waitress==1.1.0
Werkzeug==0.14.1
click==6.7
Flask==1.0.2
Flask-PyMongo==2.1.0
functools32==3.2.3.post2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pymongo==3.7.1
python-dotenv==0.9.1
waitress==1.1.0
Werkzeug==0.14.1
